move generic functions on functions to init
